refactor: kitchenpos-ui 분리
