+++ kitchenpos-ui/src/main/java/kitchenpos/product/ProductRestController.java
@@ -1,6 +1,5 @@
-package kitchenpos.product.ui;
+package kitchenpos.product;
 
-import kitchenpos.product.application.ProductService;
 import kitchenpos.product.dto.ProductRequest;
 import kitchenpos.product.dto.ProductResponse;
 import org.springframework.http.ResponseEntity;
