+++ kitchenpos-ui/src/main/resources/application.properties
