+++ kitchenpos-ui/src/main/resources/db/migration/V1__Initialize_project_tables.sql
@@ -14,7 +14,7 @@ CREATE TABLE order_line_item (
     PRIMARY KEY (seq)
 );
 
-CREATE TABLE kitchenpos.menu (
+CREATE TABLE menu (
     id BIGINT(20) NOT NULL AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
     price DECIMAL(19, 2) NOT NULL,
@@ -50,7 +50,7 @@ CREATE TABLE table_group (
     PRIMARY KEY (id)
 );
 
-CREATE TABLE kitchenpos.product (
+CREATE TABLE product (
     id BIGINT(20) NOT NULL AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
     price DECIMAL(19, 2) NOT NULL,
@@ -67,19 +67,19 @@ ALTER TABLE order_line_item
 
 ALTER TABLE order_line_item
     ADD CONSTRAINT fk_order_line_item_menu
-        FOREIGN KEY (menu_id) REFERENCES kitchenpos.menu (id);
+        FOREIGN KEY (menu_id) REFERENCES menu (id);
 
-ALTER TABLE kitchenpos.menu
+ALTER TABLE menu
     ADD CONSTRAINT fk_menu_menu_group
         FOREIGN KEY (menu_group_id) REFERENCES menu_group (id);
 
 ALTER TABLE menu_product
     ADD CONSTRAINT fk_menu_product_menu
-        FOREIGN KEY (menu_id) REFERENCES kitchenpos.menu (id);
+        FOREIGN KEY (menu_id) REFERENCES menu (id);
 
 ALTER TABLE menu_product
     ADD CONSTRAINT fk_menu_product_product
-        FOREIGN KEY (product_id) REFERENCES kitchenpos.product (id);
+        FOREIGN KEY (product_id) REFERENCES product (id);
 
 ALTER TABLE order_table
     ADD CONSTRAINT fk_order_table_table_group
