+++ kitchenpos-ui/src/test/java/kitchenpos/order/OrderAcceptanceTest.java
@@ -1,10 +1,9 @@
-package kitchenpos.order.ui;
+package kitchenpos.order;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
