+++ kitchenpos-ui/src/test/java/kitchenpos/tablegroup/TableGroupAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.ui;
+package kitchenpos.tablegroup;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -16,7 +16,7 @@ import org.springframework.http.MediaType;
 import java.util.Arrays;
 import java.util.Collections;
 
-import static kitchenpos.order.ui.OrderAcceptanceTest.주문_생성_요청;
+import static kitchenpos.order.OrderAcceptanceTest.주문_생성_요청;
 import static kitchenpos.utils.TestHelper.*;
 import static org.assertj.core.api.Assertions.assertThat;
